Move out shared parts of integration.js
